requirements.txt
alembic==1.14.1
annotated-types==0.7.0
anyio==4.6.2.post1
async-timeout==5.0.1
asyncpg==0.30.0
click==8.1.7
databases==0.9.0
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.5
fastapi-cache2==0.2.2
greenlet==3.1.1
h11==0.14.0
idna==3.10
Jinja2==3.1.6
Mako==1.3.8
MarkupSafe==3.0.2
passlib==1.7.4
pendulum==3.1.0
pydantic==2.9.2
pydantic-settings==2.7.1
pydantic_core==2.23.4
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
redis==6.0.0
setuptools==80.3.0
six==1.17.0
sniffio==1.3.1
sqladmin==0.20.1
SQLAlchemy==2.0.36
starlette==0.41.2
types-passlib==1.7.7.20241221
typing_extensions==4.12.2
tzdata==2025.2
uvicorn==0.32.0
WTForms==3.1.2
